convert: Optional object attributes
Previously we required that when converting to an object type the input
must have at least the attributes from the target type, and optionally
additional attributes that would then be discarded.

We'll now allow creating an object type with one or more optional
attributes. These are optional only for conversion purposes: converting
to an object type with optional arguments will substitute a null value
for any attribute that doesn't exist in the source object type.

This being handled only under conversion is consistent with some other
similar cases: an object type is not conforming to a type constraint if
it has attributes that are not in the type constraint, even though a
conversion to that type constraint would be accepted. Likewise, a number
doesn't conform to the string type even though a conversion is available.
This distinction is so that cty applications can potentially bring their
own separate conversion rules if they wish, ignoring the convert package
in this repository.